Enlarge the *Status columns - tinyint is not sufficient

+++ src/Dfc.CourseDirectory.Database/Pttcd/Tables/Courses.sql
@@ -1,7 +1,7 @@
 ﻿CREATE TABLE [Pttcd].[Courses]
 (
 	[CourseId] UNIQUEIDENTIFIER NOT NULL CONSTRAINT [PK_Courses] PRIMARY KEY,
-	[CourseStatus] TINYINT,
+	[CourseStatus] INT,
 	[CreatedOn] DATETIME,
 	[CreatedBy] NVARCHAR(MAX),
 	[UpdatedOn] DATETIME,
